MISC: Move System.js and Todos.js to models folder
