+++ client/react/src/models/Todo.js
@@ -1,14 +1,15 @@
 import React from 'react';
+import Box from '@mui/material/Box';
 import Card from '@mui/material/Card';
 import CardContent from '@mui/material/CardContent';
 import { CardActionArea, CardHeader } from '@mui/material';
 import Avatar from '@mui/material/Avatar';
-import GrassOutlinedIcon from '@mui/icons-material/GrassOutlined';
-import IconButton from '@mui/material/IconButton';
-import ReportGmailerrorredSharpIcon from '@mui/icons-material/ReportGmailerrorredSharp';
-import CardActions from '@mui/material/CardActions';
+// import GrassOutlinedIcon from '@mui/icons-material/GrassOutlined';
+// import IconButton from '@mui/material/IconButton';
+// import CardActions from '@mui/material/CardActions';
 import FormatListNumberedIcon from '@mui/icons-material/FormatListNumbered';
 import Typography from '@mui/material/Typography';
+// import CheckCircleOutlineIcon from '@mui/icons-material/CheckCircleOutline';
 
 const Todo = ({ todo, full }) => {
   console.log(full);
@@ -29,18 +30,20 @@ const Todo = ({ todo, full }) => {
           title={todo.created_on}
         />
         <CardContent>
-          <Typography variant="body2" color="text.secondary">
-            {todo.description}
-          </Typography>
+          <Box full>
+            <Typography variant="body2" color="text.secondary">
+              {todo.description}
+            </Typography>
+          </Box>
         </CardContent>
-        <CardActions disableSpacing>
+        {/* <CardActions disableSpacing>
           <IconButton color="info">
-            <ReportGmailerrorredSharpIcon />
+            <CheckCircleOutlineIcon />
           </IconButton>
           <IconButton color="info">
             <GrassOutlinedIcon />
           </IconButton>
-        </CardActions>
+        </CardActions> */}
       </CardActionArea>
     </Card>
   );
