+++ client/react/src/pages/Systems.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import Grid from '@mui/material/Grid';
-import System from './System';
+import System from '../models/System';
 
 
 const Systems = () => {
